test: deduplicate Darwin/Linux job-control test harness
lockedBuffer/waitForOutput/waitForProcessState and the two Ctrl-Z test bodies
were duplicated verbatim between the Darwin file and the linux-prefixed copies.
Move the shared harness into a //go:build darwin || linux helper
(integration_jobcontrol_helpers_test.go) and reduce the per-OS files to thin
OS-named entry points. `ps -o state=` behaves the same on both, so there were
no platform-specific assertions to keep.
