--- internal/sandbox/integration_jobcontrol_darwin_helpers_test.go
@@ -1,38 +1,0 @@
-//go:build darwin
-
-package sandbox
-
-import (
-	"bytes"
-	"strings"
-	"sync"
-	"time"
-)
-
-type lockedBuffer struct {
-	mu  sync.Mutex
-	buf bytes.Buffer
-}
-
-func (b *lockedBuffer) Write(p []byte) (int, error) {
-	b.mu.Lock()
-	defer b.mu.Unlock()
-	return b.buf.Write(p)
-}
-
-func (b *lockedBuffer) String() string {
-	b.mu.Lock()
-	defer b.mu.Unlock()
-	return b.buf.String()
-}
-
-func waitForOutput(output *lockedBuffer, needle string, timeout time.Duration) bool {
-	deadline := time.Now().Add(timeout)
-	for time.Now().Before(deadline) {
-		if strings.Contains(output.String(), needle) {
-			return true
-		}
-		time.Sleep(25 * time.Millisecond)
-	}
-	return false
-}
